More of VME Bridge skeleton
Filled and tested the most of program flow (I guess).

One thing not sure -- the structure for CAEN library.
Now there is a map to struct-s of one particular type of calls.
The library contains bunch of other types.
Probably, there should be a wrapper for them.
Or a switch case.