Added a stub of operations_field for future
The idea is to be able to scroll to log field and review the (DQM) output only.
Since now, Qt4 and Quick 1.0 is run -- no ScrollView, thus stopped the works.

In addition it would be nice to add the monitor-only mode for GUI.
When you connect to the operations/DQM output only.