Implemented dummy libCAENVME
It contains main batch of calls.
Always returns cvSuccess
and handles well input parameters.
For the Read cycles it returns some dummy data.

More info on VME cycles is needed
and more precise protocol for working with it:
when to use bytes, when string/integers etc.
And how to convert between them, to do checks for correct input, etc.

Implement logging for the operator with debugging mode.

	modified:   CAEN.lib/VME_operator.py
	deleted:    CAEN.lib/test/libCAENVME.so
	modified:   src/CAENVMElib.c