Changed simple_loop recorder to .so, up rakefile
simple_loop implements "record" function.
Python loads CAEN lib, sets up everything and loads simple_loop (which loads CAEN lib again).
It should be a bit more flexible -- command line arguments are parsed in Python and everything else is done in Python.
Only reading boards and writing to files is C.
In any case adding a simple main with argv parsing will be easy here.

Kind of major update of rakefile for .o and .so (and binaries later?).
The compilation writes .so lib names into resulting binary like this:

  Tag        Type                         Name/Value
 0x0000000000000001 (NEEDED)             Shared library: [shared_libs/libFoo.so]
 0x0000000000000001 (NEEDED)             Shared library: [libc.so.6]

when I want something like

  Tag        Type                         Name/Value
 0x0000000000000001 (NEEDED)             Shared library: [source.deps/libFoo.so]

-- without the path to the source, relational from the source file.
It could also be done with RPATH, but first the directory name should be altered at compilation.
It seems there is no option to change the names of input files in compilation.
And it seems like a traditional thing -- to compile everything in the same temporary directory.
Therefor now rake rules cd to source dir and compile everything from there.

Also the directory-oriented project dependencies instead of Makefile dependencies are UP.
(The .deps/ dirs and common explicit rakefile rules instead of custom Make targets for each file.)
The .deps/ contain symlinks and the targets depend on those symlinks.
If symlink is invalid rake tries to make the file as if it is in the symlink path.
The same happens for updating the target if the file pointed by symlink is newer
-- the symlink date is still old and rake thinks the dependency has not been updated.
Changed it for symlinks to .so to add actual file pointed by symlink to the dependencies.
Will propagate it to other types to -- .h are clear, should I do it for .c?
Probably yes, simply for all symlinks.
There might be some flex-generation or other files -- any dependency needs this in directory-oriented project layout.

And I need some clean-up procedure which omits the symlinks!

	modified:   rakefile
	modified:   record_run/simple_loop.c
	new file:   record_run/simple_loop.py