Cleared up project file
src/ has the structure of the languages (C or Python), the files and
directories reflect that. The C language dependencies are headers.
Therefore compilation is done by rules. But linking concrete project
files, shared and executable, is done explicitly in the rakefile. This
is the project output, to package and run in operations.

	new file:   comment.md
	modified:   project.md
	modified:   rakefile
	modified:   record_run/simple_loop.c
	modified:   record_run/simple_loop.py
	deleted:    test_caen_lib/libCAENVME.c
	deleted:    test_caen_lib/libCAENVME.deps
	modified:   utils/vmebus.py